[Todo]: Panic occurs when running on VirtualBox and actual device

+++ common/src/graphic_info.rs
@@ -7,9 +7,9 @@ pub enum PixelFormat {
 #[derive(Debug, Clone, Copy)]
 #[repr(C)]
 pub struct GraphicInfo {
-    pub resolution: (u32, u32),
+    pub resolution: (usize, usize),
     pub format: PixelFormat,
-    pub stride: u32,
+    pub stride: usize,
     pub framebuf_addr: u64,
-    pub framebuf_size: u64,
+    pub framebuf_size: usize,
 }
